Add strat9-abi as a workspace dependency and refactor bootloader integration
- Introduced the `strat9-abi` package to the workspace, enhancing modularity.
- Updated `Cargo.toml` files across various components to include `strat9-abi` as a dependency.
- Refactored bootloader code to utilize structures from `strat9-abi`, improving code clarity and reducing duplication.
- Removed redundant kernel argument structures in favor of those defined in `strat9-abi`.
- Enhanced syscall and VFS implementations to leverage new abstractions from `strat9-abi` for better maintainability.

+++ workspace/kernel/src/syscall/dispatcher.rs
@@ -93,7 +93,7 @@ fn trace_dhcp_frame(tag: &str, frame: &[u8]) {
 /// `SyscallFrame` on the current kernel stack.
 #[no_mangle]
 pub extern "C" fn __strat9_syscall_dispatch(frame: &mut SyscallFrame) -> u64 {
-    let syscall_num = frame.rax;
+    let syscall_num = frame.rax as usize;
     let arg1 = frame.rdi;
     let arg2 = frame.rsi;
     let arg3 = frame.rdx;
